+++ facebook.php
@@ -62,6 +62,16 @@ class FacebookPlugin extends Plugin {
         }
     }
 
+    private function getFacebookAccessToken($config) {
+        $url = "https://graph.facebook.com/oauth/access_token?"
+            ."client_id=".$config->get('facebook_common_settings.application_id')
+            ."&client_secret=".$config->get('facebook_common_settings.application_secret')
+            ."&grant_type=client_credentials";
+        $results = Response::get($url);
+        $json = json_decode($results);
+        return trim($json->access_token);
+    }
+
     /**
      * Add current directory to twig lookup paths.
      */
@@ -77,21 +87,19 @@ class FacebookPlugin extends Plugin {
         $twig = $this->grav['twig'];
         /** @var Data $config */
         $config = $this->mergeConfig($page, TRUE);
-
+        $accessToken = $this->getFacebookAccessToken($config);
         $filter_by_tags =
             empty($filtered_by_tags_from_page)
                 ? $config->get('facebook_page_settings.filter_by_tags')
                 : $filtered_by_tags_from_page;
+
         // Generate API url
         $url =
-            'https://graph.facebook.com/' . $config->get('facebook_page_settings.page_id')
-            . '/?fields=feed{permalink_url,created_time,link,attachments,message}&access_token='
-            . $config->get('facebook_common_settings.application_id') . '|'
-            . $config->get('facebook_common_settings.application_secret');
-            
-            
-        // http://stackoverflow.com/questions/17438847/using-facebook-graph-api-how-to-get-news-feed-with-large-picture-size-if-the-fee
+            'https://graph.facebook.com/v3.2/' . $config->get('facebook_page_settings.page_id')
+            . '/feed?access_token='. $accessToken.'&fields=permalink_url,attachments,created_time,message&limit='.$config->get('facebook_page_settings.count');
+
         $results = "";
+
         try {
             $results = Response::get($url);
             $this->parsePostResponse($results, $config, $filter_by_tags);
@@ -125,17 +133,14 @@ class FacebookPlugin extends Plugin {
         $twig = $this->grav['twig'];
         /** @var Data $config */
         $config = $this->mergeConfig($page, TRUE);
+        $accessToken = $this->getFacebookAccessToken($config);
 
         $events_page_id =
             empty($config->get('facebook_event_settings.events_page_id'))
                 ? $config->get('facebook_page_settings.page_id')
                 : $config->get('facebook_event_settings.events_page_id');
         // Generate API url
-        $url =
-            'https://graph.facebook.com/' . $events_page_id
-            . '/events?fields=cover,start_time,end_time,name,description,place&access_token='
-            . $config->get('facebook_common_settings.application_id') . '|'
-            . $config->get('facebook_common_settings.application_secret');
+        $url = 'https://graph.facebook.com/v3.2/'.$events_page_id.'/events?access_token='.$accessToken;
         $results = Response::get($url);
         $this->parseEventResponse($results, $config);
 
@@ -190,13 +195,15 @@ class FacebookPlugin extends Plugin {
     private function parsePostResponse($json, $config, $tags_string) {
         $r = array();
         $content = json_decode($json);
-        
         $count = $config->get('facebook_page_settings.count');
 
-        foreach ($content->feed->data as $val) {
+        foreach ($content->data as $val) {
             if (property_exists($val, 'message') && $this->tagsExist($tags_string, $val->message)) {
                 $created_at = $val->created_time;
                 $created_date_object = date_create($created_at);
+                if( !empty($tz = $this->grav['config']->get('system.timezone')) ) {
+                    date_timezone_set($created_date_object, timezone_open($tz));
+                }
                 $formatted_date =
                     date_format($created_date_object,
                         $config->get('facebook_page_settings.date_format'));
@@ -223,7 +230,7 @@ class FacebookPlugin extends Plugin {
                 $r[$count]['message'] = preg_split("/[\n]+/", $val->message);//nl2br($val->message);
                 $r[$count]['link'] = $val->permalink_url;
                 $this->addFeed($r);
-                
+
                 $count -= 1;
             }
         }
@@ -241,12 +248,12 @@ class FacebookPlugin extends Plugin {
                 $end_date_array = date_parse($end_at);
 
                 $start_date_array['monthName'] =
-                    date('F', mktime(0, 0, 0, $start_date_array['month'], 10));
+                    strftime ('%B', mktime(0, 0, 0, $start_date_array['month'], 10));
                 $start_date_array['dayName'] =
-                    date('l', mktime(0, 0, 0, $start_date_array['month'], $start_date_array['day'],
+                    strftime ('%A', mktime(0, 0, 0, $start_date_array['month'], $start_date_array['day'],
                         $start_date_array['year']));
                 $end_date_array['monthName'] =
-                    date('F', mktime(0, 0, 0, $end_date_array['month'], 10));
+                    strftime ('%B', mktime(0, 0, 0, $end_date_array['month'], 10));
 
                 $r[$start_at]['original_start'] = $start_at;
                 $r[$start_at]['original_end'] = $end_at;
@@ -272,7 +279,7 @@ class FacebookPlugin extends Plugin {
                 }
                 $r[$start_at]['event_link'] = $val->id;
                 $r[$start_at]['name'] = nl2br($val->name);
-                $r[$start_at]['place'] = '';
+                $r[$start_at]['place'] = array();
                 $r[$start_at]['description'] = '';
                 $r[$start_at]['cover'] = '';
 
@@ -280,9 +287,15 @@ class FacebookPlugin extends Plugin {
                     $r[$start_at]['cover'] = $val->cover;
                 }
                 if (property_exists($val, 'place')) {
-                    $r[$start_at]['place']['name'] = $val->place->name;
+                    if (property_exists($val->place, 'name')) {
+                        $r[$start_at]['place']['name'] = $val->place->name;
+                    }
                     if (property_exists($val->place, 'location')) {
-                        $r[$start_at]['place']['location'] = $val->place->location;
+                        $city = '';
+                        $country = '';
+                        if (property_exists($val->place->location, 'city')) $city = $val->place->location->city;
+                        if (property_exists($val->place->location, 'country')) $country = $val->place->location->country;
+                        $r[$start_at]['place']['location'] = $city.' '.$country;
                     }
                 }
                 if (property_exists($val, 'description')) {
@@ -300,11 +313,10 @@ class FacebookPlugin extends Plugin {
         $twig = $this->grav['twig'];
         /** @var Data $config */
         $config = $this->mergeConfig($page, TRUE);
+        $accessToken = $this->getFacebookAccessToken($config);
         // Generate API url
         $url =
-            'https://graph.facebook.com/' . $albumId . '/photos?fields=source&access_token='
-            . $config->get('facebook_common_settings.application_id') . '|'
-            . $config->get('facebook_common_settings.application_secret');
+            'https://graph.facebook.com/v3.2/' . $albumId . '/photos?fields=source&access_token='.$accessToken;
         $results = Response::get($url);
         $json = json_decode($results);
         return $json->data;